requirements.txt
discord.py==2.2.3
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.30.0
yt_dlp==2023.3.4